pyproject.toml
[project]
name = "prism"
version = "0.1.0"
description = "PRISM - Position Risk Intelligence & Swap Monitor"
requires-python = ">=3.11,<3.13"

dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.12.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "gradio>=4.0.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (import sorting)
    "B",    # flake8-bugbear (common bugs)
    "UP",   # pyupgrade (modern Python)
    "D",    # pydocstyle (docstrings)
]

ignore = [
    "E501",  # line too long
    "D100",  # missing docstring in public module
    "D104",  # missing docstring in public package
    "D203",  # incorrect-blank-line-before-class
    "D213",  # multi-line-summary-second-line
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
